Only load kubeconfig if in k8s
