+++ internal/secrets.go
@@ -24,6 +24,7 @@ import (
 	corev1 "k8s.io/api/core/v1"
 	"k8s.io/apimachinery/pkg/types"
 	"sigs.k8s.io/controller-runtime/pkg/client"
+	"sigs.k8s.io/controller-runtime/pkg/client/config"
 
 	configv1 "github.com/tetrateio/authservice-go/config/gen/go/v1"
 	oidcv1 "github.com/tetrateio/authservice-go/config/gen/go/v1/oidc"
@@ -37,8 +38,9 @@ const (
 var (
 	_ run.PreRunner = (*SecretLoader)(nil)
 
-	ErrGetSecret    = errors.New("err getting secret")
-	ErrNoSecretData = errors.New("client-secret not found in secret")
+	ErrLoadingKubeConfig = errors.New("error loading kube config")
+	ErrGetSecret         = errors.New("err getting secret")
+	ErrNoSecretData      = errors.New("client-secret not found in secret")
 )
 
 // SecretLoader is a pre-runner that loads secrets from Kubernetes and updates
@@ -51,11 +53,10 @@ type SecretLoader struct {
 
 // NewSecretLoader creates a new that loads secrets from Kubernetes and updates
 // // the configuration with the loaded data.
-func NewSecretLoader(cfg *configv1.Config, k8sClient client.Client) *SecretLoader {
+func NewSecretLoader(cfg *configv1.Config) *SecretLoader {
 	return &SecretLoader{
-		log:       Logger(Config),
-		cfg:       cfg,
-		k8sClient: k8sClient,
+		log: Logger(Config),
+		cfg: cfg,
 	}
 }
 
@@ -70,6 +71,14 @@ func (s *SecretLoader) PreRun() error {
 			oidcCfg, ok := f.Type.(*configv1.Filter_Oidc)
 			if !ok || oidcCfg.Oidc.GetClientSecretRef().GetName() == "" {
 				continue
+			}
+
+			if s.k8sClient == nil {
+				var err error
+				s.k8sClient, err = getKubeClient()
+				if err != nil {
+					return fmt.Errorf("%w: loading client secret from k8s:  %w", ErrLoadingKubeConfig, err)
+				}
 			}
 
 			errs = append(errs, s.loadClientSecretFromK8s(oidcCfg.Oidc))
@@ -109,4 +118,19 @@ func (s *SecretLoader) loadClientSecretFromK8s(cfg *oidcv1.OIDCConfig) error {
 	}
 
 	return nil
+}
+
+// getKubeClient returns a new Kubernetes client used to load secrets.
+func getKubeClient() (client.Client, error) {
+	cfg, err := config.GetConfig()
+	if err != nil {
+		return nil, fmt.Errorf("error getting kube config: %w", err)
+	}
+
+	cl, err := client.New(cfg, client.Options{})
+	if err != nil {
+		return nil, fmt.Errorf("errot creating kube client: %w", err)
+	}
+
+	return cl, nil
 }
